requirements.txt
# Core backend
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6  # for file uploads

# Embeddings and vector DB
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==0.27.9


# Data processing
pandas==2.1.1
numpy==1.26.0
pdfplumber==0.9.0
docx2txt==0.8
python-magic==0.4.27 
python-docx 
PyPDF2 
nltk

# Async handling
asyncio==3.4.3

# Optional frontend
streamlit==1.28.0

# Deployment / Docker / server
gunicorn==22.2.0
uvloop==0.17.0
jinja2==3.1.2   # needed if frontend templates used

# NLP / text processing enhancements
nltk==3.8.1
spacy==3.6.1
transformers==4.34.0
tqdm==4.66.0  # for progress bars during preprocessing
